requirement.txt
streamlit
streamlit-cropper
opencv-python-headless
numpy
scipy
Pillow
fastapi
uvicorn[standard]
python-multipart
